refactor(link): rm link mixin and fix $root events
