+++ lib/components/dropdown-item-button.js
@@ -10,13 +10,18 @@ export const props = {
 export default {
     functional: true,
     props,
-    render(h, { props, data, children }) {
+    render(h, { props, data, parent, children }) {
         return h(
             "button",
             mergeData(data, {
                 props,
                 staticClass: "dropdown-item",
-                attrs: { role: "menuitem", type: "button", disabled: props.disabled }
+                attrs: { role: "menuitem", type: "button", disabled: props.disabled },
+                on: {
+                    click(e){
+                        parent.$root.$emit("clicked::link", e)
+                    },
+                }
             }),
             children
         )
