+++ lib/components/link.js
@@ -1,61 +1,5 @@
-import { assign } from "../utils/object"
+import { assign, keys } from "../utils/object"
 import { warn, mergeData } from "../utils"
-
-function computeTag(props) {
-    return props.to && !props.disabled ? "router-link" : "a"
-}
-
-function computeRouterTag(props) {
-    if (props.tag) {
-        warn('<b-link> "tag" property is deprecated, please use "routerTag" property instead.')
-        return props.tag
-    }
-
-    return props.routerTag
-}
-
-function computeHref({ disabled, href, to }) {
-    if (disabled) return "#"
-    // If href explicitly provided
-    if (href) return href
-    // Fallback to `to` prop
-    if (to && typeof to === "string") return to
-}
-
-function computeRel({ target, rel }) {
-    if (target === "_blank" && rel === null) {
-        return "noopener"
-    }
-    return rel || null
-}
-
-function computeClass(props) {
-    return [
-        props.active ? (props.exact ? props.exactActiveClass : props.activeClass) : null,
-        props.disabled ? "disabled" : null
-    ]
-}
-
-function clickHandlerFactory({ disabled, tag, href, parent }) {
-    return function onClick(e) {
-        const isRouterLink = tag === "router-link"
-
-        if (disabled) {
-            e.stopPropagation()
-        } else {
-            parent.$root.$emit("clicked::link")
-
-            if (isRouterLink) {
-                e.target.__vue__.$emit("click", e)
-            }
-        }
-
-        if (!isRouterLink && href === "#") {
-            // stop scroll-to-top behavior
-            e.preventDefault()
-        }
-    }
-}
 
 export const anchorProps = {
     href: {
@@ -121,12 +65,85 @@ export const routerProps = {
 
 export const props = assign({}, routerProps, anchorProps)
 
+export const computed = {
+    linkProps() {
+        let linkProps = {},
+            propKeys = keys(props)
+
+        for (let i = 0; i < propKeys.length; i++) {
+            const prop = propKeys[i]
+            // Computed Vue getters are bound to the instance.
+            linkProps[prop] = this[prop]
+        }
+
+        return linkProps
+    }
+}
+
+function computeTag(props) {
+    return props.to && !props.disabled ? "router-link" : "a"
+}
+
+function computeRouterTag(props) {
+    if (props.tag) {
+        warn('<b-link> "tag" property is deprecated, please use "routerTag" property instead.')
+        return props.tag
+    }
+
+    return props.routerTag
+}
+
+function computeHref({ disabled, href, to }) {
+    if (disabled) return "#"
+    // If href explicitly provided
+    if (href) return href
+    // Fallback to `to` prop
+    if (to && typeof to === "string") return to
+}
+
+function computeRel({ target, rel }) {
+    if (target === "_blank" && rel === null) {
+        return "noopener"
+    }
+    return rel || null
+}
+
+function computeClass(props) {
+    return [
+        props.active ? (props.exact ? props.exactActiveClass : props.activeClass) : null,
+        props.disabled ? "disabled" : null
+    ]
+}
+
+function clickHandlerFactory({ disabled, tag, href, parent }) {
+    return function onClick(e) {
+        const isRouterLink = tag === "router-link"
+
+        if (disabled) {
+            e.stopPropagation()
+        } else {
+            parent.$root.$emit("clicked::link", e)
+
+            if (isRouterLink) {
+                e.target.__vue__.$emit("click", e)
+            }
+        }
+
+        if (!isRouterLink && href === "#") {
+            // stop scroll-to-top behavior
+            e.preventDefault()
+        }
+    }
+}
+
 export default {
     functional: true,
     props: props,
-    render(h, { props, data, children, parent }) {
-        const [tag, rel, href] = [computeTag(props), computeRel(props), computeHref(props)]
-        const onClick = clickHandlerFactory({ tag, href, disabled: props.disabled, parent })
+    render(h, { props, data, parent, children }) {
+        const tag = computeTag(props),
+            rel = computeRel(props),
+            href = computeHref(props),
+            onClick = clickHandlerFactory({ tag, href, disabled: props.disabled, parent })
 
         const componentData = {
             class: computeClass(props),
