+++ lib/mixins/dropdown.js
@@ -1,5 +1,7 @@
 import clickOut from './clickout';
+import listenOnRootMixin from './listen-on-root'
 import { from as arrayFrom } from '../utils/array'
+import Vue from 'vue'
 
 // Determine if an HTML element is visible - Faster than CSS check
 function isVisible(el) {
@@ -17,6 +19,7 @@ const HEADER_SELECTOR = '.dropdown-header';
 const ALL_SELECTOR = [ITEM_SELECTOR, HEADER_SELECTOR].join(',');
 
 export default {
+    mixins: [listenOnRootMixin],
     props: {
         id: {
             type: String
@@ -42,6 +45,19 @@ export default {
         return {
             visible: false
         };
+    },
+    created() {
+        const listener = el => {
+            if (el !== this) {
+                this.visible = false;
+            }
+        };
+
+        // To keep one dropdown opened on page
+        this.listenOnRoot('shown::dropdown', listener);
+
+        // Hide when clicked on links
+        this.listenOnRoot('clicked::link', listener);
     },
     mounted: clickOut.mounted,
     destroyed: clickOut.destroyed,
