+++ lib/mixins/index.js
@@ -4,7 +4,6 @@ import formMixin from "./form"
 import formCheckBoxMixin from "./form-checkbox"
 import formCustomMixin from "./form-custom"
 import formOptionsMixin from "./form-options"
-import linkMixin from "./link"
 import listenOnRootMixin from "./listen-on-root"
 import popoverMixin from "./popover"
 
@@ -15,7 +14,6 @@ export {
     formCheckBoxMixin,
     formCustomMixin,
     formOptionsMixin,
-    linkMixin,
     listenOnRootMixin,
     popoverMixin
 }
